Billboard.fx and Renderbillboard is no longer used and was removed. Fixed some warnings.

+++ NDYGFX/Source/CascadedShadowMap.cpp
@@ -105,7 +105,7 @@ void CascadedShadowMap::CalcShadowMapMatrices(D3DXVECTOR3 sunLight, Camera* cam,
 	
 	D3DXVECTOR3 tmpNearPlanePoint = minValue;
 	//Set the near plane to be closer to the light to include more potential occluders.
-	tmpNearPlanePoint -= sunLight * this->params.ShadowFit * (i + 1); //TILLMAN** sätta near plane till samma för alla.
+	tmpNearPlanePoint -= sunLight * (float)(this->params.ShadowFit * (i + 1)); //TILLMAN** sätta near plane till samma för alla.
 	float nearPlane = tmpNearPlanePoint.z; 
 	float farPlane = maxValue.z;
